+++ Editor/XRReferenceImageLibraryBuildProcessor.cs
@@ -6,11 +6,9 @@ using UnityEngine.XR.ARSubsystems;
 
 namespace UnityEditor.XR.ARSubsystems
 {
-    class XRReferenceImageLibraryBuildProcessor : IPreprocessBuildWithReport, IPostprocessBuildWithReport
+    static class XRReferenceImageLibraryBuildProcessor
     {
-        public int callbackOrder => -1000;
-
-        static void ClearDataStore()
+        public static void ClearDataStore()
         {
             foreach (var library in XRReferenceImageLibrary.All())
             {
@@ -18,9 +16,21 @@ namespace UnityEditor.XR.ARSubsystems
             }
         }
 
-        public void OnPostprocessBuild(BuildReport report) => ClearDataStore();
+        class Preprocessor : IPreprocessBuildWithReport
+        {
+            // Should come before most other things
+            public int callbackOrder => -1000;
 
-        // Clear all native data and let each provider decide whether to populate it or not.
-        public void OnPreprocessBuild(BuildReport report) => ClearDataStore();
+            // Clear all native data and let each provider decide whether to populate it or not.
+            public void OnPreprocessBuild(BuildReport report) => ClearDataStore();
+        }
+
+        class Postprocessor : IPostprocessBuildWithReport
+        {
+            // Should come after most other things
+            public int callbackOrder => 1000;
+
+            public void OnPostprocessBuild(BuildReport report) => ClearDataStore();
+        }
     }
 }